Module rustc_passes::check_attr

This module implements some validity checks for attributes. In particular it verifies that #[inline] and #[repr] attributes are attached to items that actually support them and if there are conflicts between multiple such attributes attached to the same item.

Structs

CheckAttrVisitor ğŸ”’

Enums

ItemLike ğŸ”’

Functions

check_duplicates ğŸ”’

check_invalid_crate_level_attr ğŸ”’

check_mod_attrs ğŸ”’

check_non_exported_macro_for_invalid_attrs ğŸ”’

is_c_like_enum ğŸ”’

provide ğŸ”’

target_from_impl_item ğŸ”’