Struct rustc_passes::dead::MarkSymbolVisitor

struct MarkSymbolVisitor<'tcx> {
    worklist: Vec<LocalDefId>,
    tcx: TyCtxt<'tcx>,
    maybe_typeck_results: Option<&'tcx TypeckResults<'tcx>>,
    live_symbols: FxHashSet<LocalDefId>,
    repr_has_repr_c: bool,
    repr_has_repr_simd: bool,
    in_pat: bool,
    ignore_variant_stack: Vec<DefId>,
    struct_constructors: FxHashMap<LocalDefId, LocalDefId>,
    ignored_derived_traits: FxHashMap<LocalDefId, Vec<(DefId, DefId)>>,
}

Fields
worklist: Vec<LocalDefId>tcx: TyCtxt<'tcx>maybe_typeck_results: Option<&'tcx TypeckResults<'tcx>>live_symbols: FxHashSet<LocalDefId>repr_has_repr_c: boolrepr_has_repr_simd: boolin_pat: boolignore_variant_stack: Vec<DefId>struct_constructors: FxHashMap<LocalDefId, LocalDefId>ignored_derived_traits: FxHashMap<LocalDefId, Vec<(DefId, DefId)>>
Implementations

impl<'tcx> MarkSymbolVisitor<'tcx>

fn typeck_results(&self) -> &'tcx TypeckResults<'tcx>

Gets the type-checking results for the current body. As this will ICE if called outside bodies, only call when working with Expr or Pat nodes (they are guaranteed to be found only in bodies).

fn check_def_id(&mut self, def_id: DefId)

fn insert_def_id(&mut self, def_id: DefId)

fn handle_res(&mut self, res: Res)

fn lookup_and_handle_method(&mut self, id: HirId)

fn handle_field_access(&mut self, lhs: &Expr<'_>, hir_id: HirId)

fn handle_assign(&mut self, expr: &'tcx Expr<'tcx>)

fn check_for_self_assign(&mut self, assign: &'tcx Expr<'tcx>)

fn handle_field_pattern_match(
&mut self,
lhs: &Pat<'_>,
res: Res,
pats: &[PatField<'_>]
)

fn handle_tuple_field_pattern_match(
&mut self,
lhs: &Pat<'_>,
res: Res,
pats: &[Pat<'_>],
dotdot: DotDotPos
)

fn mark_live_symbols(&mut self)

fn should_ignore_item(&mut self, def_id: DefId) -> bool

Automatically generated items marked with rustc_trivial_field_reads will be ignored for the purposes of dead code analysis (see PR #85200 for discussion).

fn visit_node(&mut self, node: Node<'tcx>)

fn mark_as_used_if_union(&mut self, adt: AdtDef<'tcx>, fields: &[ExprField<'_>])