Struct rustc_passes::hir_stats::StatCollector

struct StatCollector<'k> {
    krate: Option<Map<'k>>,
    nodes: FxHashMap<&'static str, Node>,
    seen: FxHashSet<Id>,
}

This type measures the size of AST and HIR nodes, by implementing the AST and HIR Visitor traits. But we don’t measure every visited type because that could cause double counting.

For example, ast::Visitor has visit_ident, but Idents are always stored inline within other AST nodes, so we don’t implement visit_ident here. In contrast, we do implement visit_expr because ast::Expr is always stored as P<ast::Expr>, and every such expression should be measured separately.

In general, a visit_foo method should be implemented here if the corresponding Foo type is always stored on its own, e.g.: P<Foo>, Box<Foo>, Vec<Foo>, Box<[Foo]>.

There are some types in the AST and HIR tree that the visitors do not have a visit_* method for, and so we cannot measure these, which is unfortunate.

Fields
krate: Option<Map<'k>>nodes: FxHashMap<&'static str, Node>seen: FxHashSet<Id>
Implementations

impl<'k> StatCollector<'k>

fn record<T>(&mut self, label: &'static str, id: Id, val: &T)

fn record_variant<T>(
&mut self,
label1: &'static str,
label2: &'static str,
id: Id,
val: &T
)

fn record_inner<T>(
&mut self,
label1: &'static str,
label2: Option<&'static str>,
id: Id,
val: &T
)

fn print(&self, title: &str, prefix: &str)