Struct rustc_passes::liveness::rwu_table::RWUTable

pub(super) struct RWUTable {
    live_nodes: usize,
    vars: usize,
    words: Vec<u8>,
    live_node_words: usize,
}

Conceptually, this is like a Vec<Vec<RWU>>. But the number of RWU`s can get very large, so it uses a more compact representation.

Fields
live_nodes: usize

Total number of live nodes.

vars: usize

Total number of variables.

words: Vec<u8>

A compressed representation of RWUs.

Each word represents 2 different RWUs packed together. Each packed RWU is stored in 4 bits: a reader bit, a writer bit, a used bit and a padding bit.

The data for each live node is contiguous and starts at a word boundary, so there might be an unused space left.

live_node_words: usize

Number of words per each live node.

Implementations

impl RWUTable

const RWU_READER: u8 = 1u8

const RWU_WRITER: u8 = 2u8

const RWU_USED: u8 = 4u8

const RWU_MASK: u8 = 15u8

const RWU_BITS: usize = 4usize

Size of packed RWU in bits.

const WORD_BITS: usize = 8usize

Size of a word in bits.

const WORD_RWU_COUNT: usize = 2usize

Number of packed RWUs that fit into a single word.

pub(super) fn new(live_nodes: usize, vars: usize) -> RWUTable

fn word_and_shift(&self, ln: LiveNode, var: Variable) -> (usize, u32)

fn pick2_rows_mut(&mut self, a: LiveNode, b: LiveNode) -> (&mut [u8], &mut [u8])

pub(super) fn copy(&mut self, dst: LiveNode, src: LiveNode)

pub(super) fn union(&mut self, dst: LiveNode, src: LiveNode) -> bool

Sets dst to the union of dst and src, returns true if dst was changed.

pub(super) fn get_reader(&self, ln: LiveNode, var: Variable) -> bool

pub(super) fn get_writer(&self, ln: LiveNode, var: Variable) -> bool

pub(super) fn get_used(&self, ln: LiveNode, var: Variable) -> bool

pub(super) fn get(&self, ln: LiveNode, var: Variable) -> RWU

pub(super) fn set(&mut self, ln: LiveNode, var: Variable, rwu: RWU)