Struct rustc_passes::liveness::Liveness

struct Liveness<'a, 'tcx> {
    ir: &'a mut IrMaps<'tcx>,
    typeck_results: &'a TypeckResults<'tcx>,
    param_env: ParamEnv<'tcx>,
    closure_min_captures: Option<&'tcx RootVariableMinCaptureList<'tcx>>,
    successors: IndexVec<LiveNode, Option<LiveNode>>,
    rwu_table: RWUTable,
    closure_ln: LiveNode,
    exit_ln: LiveNode,
    break_ln: HirIdMap<LiveNode>,
    cont_ln: HirIdMap<LiveNode>,
}

Fields
ir: &'a mut IrMaps<'tcx>typeck_results: &'a TypeckResults<'tcx>param_env: ParamEnv<'tcx>closure_min_captures: Option<&'tcx RootVariableMinCaptureList<'tcx>>successors: IndexVec<LiveNode, Option<LiveNode>>rwu_table: RWUTableclosure_ln: LiveNode

A live node representing a point of execution before closure entry & after closure exit. Used to calculate liveness of captured variables through calls to the same closure. Used for Fn & FnMut closures only.

exit_ln: LiveNode

A live node representing every ‘exit’ from the function, whether it be by explicit return, panic, or other means.

break_ln: HirIdMap<LiveNode>cont_ln: HirIdMap<LiveNode>
Implementations

impl<'a, 'tcx> Liveness<'a, 'tcx>

fn new(ir: &'a mut IrMaps<'tcx>, body_owner: LocalDefId) -> Liveness<'a, 'tcx>

fn live_node(&self, hir_id: HirId, span: Span) -> LiveNode

fn variable(&self, hir_id: HirId, span: Span) -> Variable

fn define_bindings_in_pat(&mut self, pat: &Pat<'_>, succ: LiveNode) -> LiveNode

fn live_on_entry(&self, ln: LiveNode, var: Variable) -> bool

fn live_on_exit(&self, ln: LiveNode, var: Variable) -> bool

fn used_on_entry(&self, ln: LiveNode, var: Variable) -> bool

fn assigned_on_entry(&self, ln: LiveNode, var: Variable) -> bool

fn assigned_on_exit(&self, ln: LiveNode, var: Variable) -> bool

fn write_vars<F>(&self, wr: &mut dyn Write, test: F) -> Result<()>where
F: FnMut(Variable) -> bool,

fn ln_str(&self, ln: LiveNode) -> String

fn log_liveness(&self, entry_ln: LiveNode, hir_id: HirId)

fn init_empty(&mut self, ln: LiveNode, succ_ln: LiveNode)

fn init_from_succ(&mut self, ln: LiveNode, succ_ln: LiveNode)

fn merge_from_succ(&mut self, ln: LiveNode, succ_ln: LiveNode) -> bool

fn define(&mut self, writer: LiveNode, var: Variable)

fn acc(&mut self, ln: LiveNode, var: Variable, acc: u32)

fn compute(&mut self, body: &Body<'_>, hir_id: HirId) -> LiveNode

fn propagate_through_block(&mut self, blk: &Block<'_>, succ: LiveNode) -> LiveNode

fn propagate_through_stmt(&mut self, stmt: &Stmt<'_>, succ: LiveNode) -> LiveNode

fn propagate_through_exprs(
&mut self,
exprs: &[Expr<'_>],
succ: LiveNode
) -> LiveNode

fn propagate_through_opt_expr(
&mut self,
opt_expr: Option<&Expr<'_>>,
succ: LiveNode
) -> LiveNode

fn propagate_through_expr(&mut self, expr: &Expr<'_>, succ: LiveNode) -> LiveNode

fn propagate_through_place_components(
&mut self,
expr: &Expr<'_>,
succ: LiveNode
) -> LiveNode

fn write_place(&mut self, expr: &Expr<'_>, succ: LiveNode, acc: u32) -> LiveNode

fn access_var(
&mut self,
hir_id: HirId,
var_hid: HirId,
succ: LiveNode,
acc: u32,
span: Span
) -> LiveNode

fn access_path(
&mut self,
hir_id: HirId,
path: &Path<'_>,
succ: LiveNode,
acc: u32
) -> LiveNode

fn propagate_through_loop(
&mut self,
expr: &Expr<'_>,
body: &Block<'_>,
succ: LiveNode
) -> LiveNode

fn check_is_ty_uninhabited(&mut self, expr: &Expr<'_>, succ: LiveNode) -> LiveNode

fn warn_about_unreachable(
&mut self,
orig_span: Span,
orig_ty: Ty<'tcx>,
expr_span: Span,
expr_id: HirId,
descr: &str
)

impl<'tcx> Liveness<'_, 'tcx>

fn check_place(&mut self, expr: &'tcx Expr<'tcx>)

fn should_warn(&self, var: Variable) -> Option<String>

fn warn_about_unused_upvars(&self, entry_ln: LiveNode)

fn warn_about_unused_args(&self, body: &Body<'_>, entry_ln: LiveNode)

fn check_unused_vars_in_pat(
&self,
pat: &Pat<'_>,
entry_ln: Option<LiveNode>,
opt_body: Option<&Body<'_>>,
on_used_on_entry: impl Fn(Vec<Span>, HirId, LiveNode, Variable)
)

fn report_unused(
&self,
hir_ids_and_spans: Vec<(HirId, Span, Span)>,
ln: LiveNode,
var: Variable,
can_remove: bool,
pat: &Pat<'_>,
opt_body: Option<&Body<'_>>
)

fn has_added_lit_match_name_span(
&self,
name: &str,
opt_body: Option<&Body<'_>>,
err: &mut DiagnosticBuilder<'_, ()>
) -> bool

fn warn_about_dead_assign(
&self,
spans: Vec<Span>,
hir_id: HirId,
ln: LiveNode,
var: Variable
)

fn report_unused_assign(
&self,
hir_id: HirId,
spans: Vec<Span>,
var: Variable,
message: impl Fn(&str) -> String
)