Struct rustc_passes::check_attr::CheckAttrVisitor

struct CheckAttrVisitor<'tcx> {
    tcx: TyCtxt<'tcx>,
}

Fields
tcx: TyCtxt<'tcx>
Implementations

impl CheckAttrVisitor<'_>

fn check_attributes(
&self,
hir_id: HirId,
span: Span,
target: Target,
item: Option<ItemLike<'_>>
)

Checks any attribute.

fn inline_attr_str_error_with_macro_def(
&self,
hir_id: HirId,
attr: &Attribute,
sym: &str
)

fn inline_attr_str_error_without_macro_def(
&self,
hir_id: HirId,
attr: &Attribute,
sym: &str
)

fn check_inline(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if an #[inline] is applied to a function or a closure. Returns true if valid.

fn check_no_coverage(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if a #[no_coverage] is applied directly to a function

fn check_generic_attr(
&self,
hir_id: HirId,
attr: &Attribute,
target: Target,
allowed_targets: &[Target]
)

fn check_naked(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if #[naked] is applied to a function definition.

fn check_cmse_nonsecure_entry(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if #[cmse_nonsecure_entry] is applied to a function definition.

fn check_object_lifetime_default(&self, hir_id: HirId)

Debugging aid for object_lifetime_default query.

fn check_collapse_debuginfo(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if #[collapse_debuginfo] is applied to a macro.

fn check_track_caller(
&self,
hir_id: HirId,
attr_span: Span,
attrs: &[Attribute],
span: Span,
target: Target
) -> bool

Checks if a #[track_caller] is applied to a non-naked function. Returns true if valid.

fn check_non_exhaustive(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if the #[non_exhaustive] attribute on an item is valid. Returns true if valid.

fn check_marker(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if the #[marker] attribute on an item is valid. Returns true if valid.

fn check_rustc_must_implement_one_of(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if the #[rustc_must_implement_one_of] attribute on a target is valid. Returns true if valid.

fn check_target_feature(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if the #[target_feature] attribute on item is valid. Returns true if valid.

fn check_thread_local(&self, attr: &Attribute, span: Span, target: Target) -> bool

Checks if the #[thread_local] attribute on item is valid. Returns true if valid.

fn doc_attr_str_error(&self, meta: &NestedMetaItem, attr_name: &str)

fn check_doc_alias_value(
&self,
meta: &NestedMetaItem,
doc_alias: Symbol,
hir_id: HirId,
target: Target,
is_list: bool,
aliases: &mut FxHashMap<String, Span>
) -> bool

fn check_doc_alias(
&self,
meta: &NestedMetaItem,
hir_id: HirId,
target: Target,
aliases: &mut FxHashMap<String, Span>
) -> bool

fn check_doc_keyword(&self, meta: &NestedMetaItem, hir_id: HirId) -> bool

fn check_doc_fake_variadic(&self, meta: &NestedMetaItem, hir_id: HirId) -> bool

fn check_doc_inline(
&self,
attr: &Attribute,
meta: &NestedMetaItem,
hir_id: HirId,
target: Target,
specified_inline: &mut Option<(bool, Span)>
) -> bool

Checks #[doc(inline)]/#[doc(no_inline)] attributes. Returns true if valid.

A doc inlining attribute is invalid if it is applied to a non-use item, or if there are conflicting attributes for one item.

specified_inline is used to keep track of whether we have already seen an inlining attribute for this item. If so, specified_inline holds the value and the span of the first inline/no_inline attribute.

fn check_attr_not_crate_level(
&self,
meta: &NestedMetaItem,
hir_id: HirId,
attr_name: &str
) -> bool

Checks that an attribute is not used at the crate level. Returns true if valid.

fn check_attr_crate_level(
&self,
attr: &Attribute,
meta: &NestedMetaItem,
hir_id: HirId
) -> bool

Checks that an attribute is used at the crate level. Returns true if valid.

fn check_test_attr(&self, meta: &NestedMetaItem, hir_id: HirId) -> bool

Checks that doc(test(...)) attribute contains only valid attributes. Returns true if valid.

fn check_doc_attrs(
&self,
attr: &Attribute,
hir_id: HirId,
target: Target,
specified_inline: &mut Option<(bool, Span)>,
aliases: &mut FxHashMap<String, Span>
) -> bool

Runs various checks on #[doc] attributes. Returns true if valid.

specified_inline should be initialized to None and kept for the scope of one item. Read the documentation of check_doc_inline for more information.

fn check_pass_by_value(&self, attr: &Attribute, span: Span, target: Target) -> bool

Warns against some misuses of #[pass_by_value]

fn check_allow_incoherent_impl(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

fn check_has_incoherent_inherent_impls(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

fn check_must_use(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Warns against some misuses of #[must_use]

fn check_must_not_suspend(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if #[must_not_suspend] is applied to a function. Returns true if valid.

fn check_cold(&self, hir_id: HirId, attr: &Attribute, span: Span, target: Target)

Checks if #[cold] is applied to a non-function. Returns true if valid.

fn check_link(&self, hir_id: HirId, attr: &Attribute, span: Span, target: Target)

Checks if #[link] is applied to an item other than a foreign module.

fn check_link_name(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
)

Checks if #[link_name] is applied to an item other than a foreign function or static.

fn check_no_link(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if #[no_link] is applied to an extern crate. Returns true if valid.

fn is_impl_item(&self, hir_id: HirId) -> bool

fn check_export_name(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks if #[export_name] is applied to a function or static. Returns true if valid.

fn check_rustc_layout_scalar_valid_range(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

fn check_rustc_legacy_const_generics(
&self,
attr: &Attribute,
span: Span,
target: Target,
item: Option<ItemLike<'_>>
) -> bool

Checks if #[rustc_legacy_const_generics] is applied to a function and has a valid argument.

fn check_applied_to_fn_or_method(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Helper function for checking that the provided attribute is only applied to a function or method.

fn check_rustc_lint_query_instability(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks that the #[rustc_lint_query_instability] attribute is only applied to a function or method.

fn check_rustc_lint_diagnostics(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks that the #[rustc_lint_diagnostics] attribute is only applied to a function or method.

fn check_rustc_lint_opt_ty(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks that the #[rustc_lint_opt_ty] attribute is only applied to a struct.

fn check_rustc_lint_opt_deny_field_access(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Checks that the #[rustc_lint_opt_deny_field_access] attribute is only applied to a field.

fn check_rustc_dirty_clean(&self, attr: &Attribute) -> bool

Checks that the dep-graph debugging attributes are only present when the query-dep-graph option is passed to the compiler.

fn check_link_section(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
)

Checks if #[link_section] is applied to a function or static.

fn check_no_mangle(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
)

Checks if #[no_mangle] is applied to a function or static.

fn check_repr(
&self,
attrs: &[Attribute],
span: Span,
target: Target,
item: Option<ItemLike<'_>>,
hir_id: HirId
)

Checks if the #[repr] attributes on item are valid.

fn check_used(&self, attrs: &[Attribute], target: Target)

fn check_allow_internal_unstable(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target,
attrs: &[Attribute]
) -> bool

Outputs an error for #[allow_internal_unstable] which can only be applied to macros. (Allows proc_macro functions)

fn check_debugger_visualizer(&self, attr: &Attribute, target: Target) -> bool

Checks if the items on the #[debugger_visualizer] attribute are valid.

fn check_rustc_allow_const_fn_unstable(
&self,
hir_id: HirId,
attr: &Attribute,
span: Span,
target: Target
) -> bool

Outputs an error for #[allow_internal_unstable] which can only be applied to macros. (Allows proc_macro functions)

fn check_rustc_std_internal_symbol(
&self,
attr: &Attribute,
span: Span,
target: Target
) -> bool

fn check_const_trait(&self, attr: &Attribute, _span: Span, target: Target) -> bool

#[const_trait] only applies to traits.

fn check_stability_promotable(
&self,
attr: &Attribute,
_span: Span,
target: Target
) -> bool

fn check_link_ordinal(&self, attr: &Attribute, _span: Span, target: Target) -> bool

fn check_deprecated(
&self,
hir_id: HirId,
attr: &Attribute,
_span: Span,
target: Target
)

fn check_macro_use(&self, hir_id: HirId, attr: &Attribute, target: Target)

fn check_macro_export(&self, hir_id: HirId, attr: &Attribute, target: Target)

fn check_plugin_registrar(&self, hir_id: HirId, attr: &Attribute, target: Target)

fn check_unused_attribute(&self, hir_id: HirId, attr: &Attribute)