Module rustc_passes::diagnostic_items

Detecting diagnostic items.

Diagnostic items are items that are not language-inherent, but can reasonably be expected to exist for diagnostic purposes. This allows diagnostic authors to refer to specific items directly, without having to guess module paths and crates. Examples are:

Traits like Debug, that have no bearing on language semantics

Compiler internal types like Ty and TyCtxt

Functions

all_diagnostic_items 🔒

Traverse and collect all the diagnostic items in all crates.

collect_item 🔒

diagnostic_items 🔒

Traverse and collect the diagnostic items in the current

extract 🔒

Extract the first rustc_diagnostic_item = "$name" out of a list of attributes.

observe_item 🔒

provide

report_duplicate_item 🔒